Move TargetIdAuthorization, SecondaryKeyWithAuth, and CreateChildIdentityWithAuth to primitives.
