+++ pallets/common/src/traits/identity.rs
@@ -18,7 +18,6 @@ use frame_support::dispatch::{DispatchError, GetDispatchInfo, PostDispatchInfo, 
 use frame_support::traits::{Currency, EnsureOrigin, Get, GetCallMetadata};
 use frame_support::{decl_event, Parameter};
 use scale_info::TypeInfo;
-use sp_core::H512;
 use sp_runtime::traits::{Dispatchable, IdentifyAccount, Member, Verify};
 use sp_std::vec::Vec;
 
@@ -26,54 +25,10 @@ use polymesh_primitives::identity::limits::{
     MAX_ASSETS, MAX_EXTRINSICS, MAX_PALLETS, MAX_PORTFOLIOS,
 };
 use polymesh_primitives::{
-    protocol_fee::ChargeProtocolFee, secondary_key::SecondaryKey, traits::group::GroupTrait,
-    traits::CddAndFeeDetails, AuthorizationData, Balance, CustomClaimTypeId, IdentityClaim,
-    IdentityId, Permissions, Ticker,
+    identity::SecondaryKeyWithAuth, protocol_fee::ChargeProtocolFee, secondary_key::SecondaryKey,
+    traits::group::GroupTrait, traits::CddAndFeeDetails, AuthorizationData, Balance,
+    CustomClaimTypeId, IdentityClaim, IdentityId, Permissions, Ticker,
 };
-
-pub type AuthorizationNonce = u64;
-
-/// It represents an authorization that any account could sign to allow operations related with a
-/// target identity.
-///
-/// # Safety
-///
-/// Please note, that `nonce` has been added to avoid **replay attack** and it should be the current
-/// value of nonce of primary key of `target_id`. See `System::account_nonce`.
-/// In this way, the authorization is delimited to an specific transaction (usually the next one)
-/// of primary key of target identity.
-#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug)]
-pub struct TargetIdAuthorization<Moment> {
-    /// Target identity which is authorized to make an operation.
-    pub target_id: IdentityId,
-    /// It HAS TO be `target_id` authorization nonce: See `Identity::offchain_authorization_nonce`
-    pub nonce: AuthorizationNonce,
-    pub expires_at: Moment,
-}
-
-/// Secondary key with authorization of that secondary key (off-chain operation) to be added
-/// to an identity.
-///
-/// `auth_signature` is the signature, generated by secondary key, of `TargetIdAuthorization`.
-#[derive(Encode, Decode, TypeInfo, Clone, PartialEq, Eq, Debug)]
-pub struct SecondaryKeyWithAuth<AccountId> {
-    /// Secondary key to be added.
-    pub secondary_key: SecondaryKey<AccountId>,
-    /// Off-chain authorization signature.
-    pub auth_signature: H512,
-}
-
-/// Create a child identity using `key` as the primary key of the new child identity.
-///
-/// The `key` needs to sign (off-chain) an authorization.
-#[derive(Encode, Decode, TypeInfo, Clone, PartialEq, Eq, Debug)]
-pub struct CreateChildIdentityWithAuth<AccountId> {
-    /// The key to be used as the primary key of a new child identity.
-    pub key: AccountId,
-    /// Off-chain authorization signature.
-    /// The signature is generated by `key` signing of `TargetIdAuthorization`.
-    pub auth_signature: H512,
-}
 
 pub trait WeightInfo {
     fn create_child_identity() -> Weight;
