+++ primitives/src/identity.rs
@@ -1,3 +1,4 @@
+#![allow(missing_docs)]
 // This file is part of the Polymesh distribution (https://github.com/PolymeshAssociation/Polymesh).
 // Copyright (c) 2020 Polymesh Association
 
@@ -18,6 +19,9 @@ use sp_runtime::{Deserialize, Serialize};
 
 use codec::{Decode, Encode};
 use scale_info::TypeInfo;
+use sp_core::H512;
+
+use crate::{secondary_key::SecondaryKey, IdentityId};
 
 #[cfg(feature = "running-ci")]
 /// Defines the constants for the identity pallet.
@@ -75,4 +79,48 @@ impl<AccountId> DidRecord<AccountId> {
             primary_key: Some(primary_key),
         }
     }
+}
+
+pub type AuthorizationNonce = u64;
+
+/// It represents an authorization that any account could sign to allow operations related with a
+/// target identity.
+///
+/// # Safety
+///
+/// Please note, that `nonce` has been added to avoid **replay attack** and it should be the current
+/// value of nonce of primary key of `target_id`. See `System::account_nonce`.
+/// In this way, the authorization is delimited to an specific transaction (usually the next one)
+/// of primary key of target identity.
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug)]
+pub struct TargetIdAuthorization<Moment> {
+    /// Target identity which is authorized to make an operation.
+    pub target_id: IdentityId,
+    /// It HAS TO be `target_id` authorization nonce: See `Identity::offchain_authorization_nonce`
+    pub nonce: AuthorizationNonce,
+    pub expires_at: Moment,
+}
+
+/// Secondary key with authorization of that secondary key (off-chain operation) to be added
+/// to an identity.
+///
+/// `auth_signature` is the signature, generated by secondary key, of `TargetIdAuthorization`.
+#[derive(Encode, Decode, TypeInfo, Clone, PartialEq, Eq, Debug)]
+pub struct SecondaryKeyWithAuth<AccountId> {
+    /// Secondary key to be added.
+    pub secondary_key: SecondaryKey<AccountId>,
+    /// Off-chain authorization signature.
+    pub auth_signature: H512,
+}
+
+/// Create a child identity using `key` as the primary key of the new child identity.
+///
+/// The `key` needs to sign (off-chain) an authorization.
+#[derive(Encode, Decode, TypeInfo, Clone, PartialEq, Eq, Debug)]
+pub struct CreateChildIdentityWithAuth<AccountId> {
+    /// The key to be used as the primary key of a new child identity.
+    pub key: AccountId,
+    /// Off-chain authorization signature.
+    /// The signature is generated by `key` signing of `TargetIdAuthorization`.
+    pub auth_signature: H512,
 }
